Select spike times and amplitudes more efficiently
When the number of spikes and templates is large, it becomes very expensive to select values from the vectors spikeTimes_seconds and templateAmplitudes iteratively one template at a time. By moving this step outside of the main loop, and using accumarray to group the values by template ID, a speedup of around 2x can be achieved for the runAllQualityMetrics function (with 'extractRaw' disabled).